refactor: 컴포넌트명 명확하게 변경 (#43)
