+++ src/components/layout/header/HeaderNavigation.tsx
@@ -1,21 +1,21 @@
-import HeaderIcon from '@/components/common/HeaderIcon';
+import NavIcon from '@/components/common/NavIcon';
 import { NationIcon, CartIcon, UserIcon } from '@/assets/icons';
 import { ROUTER_PATH, PATH_NAME } from '@/constants';
 
-const NavIcon = () => {
+const HeaderNavigation = () => {
   return (
     <div className="gap-xl flex">
-      <HeaderIcon
+      <NavIcon
         href={ROUTER_PATH.COUNTRY_SETTING}
         icon={<NationIcon />}
         label={PATH_NAME.COUNTRY_SETTING}
       />
-      <HeaderIcon
+      <NavIcon
         href={ROUTER_PATH.CART}
         icon={<CartIcon />}
         label={PATH_NAME.CART}
       />
-      <HeaderIcon
+      <NavIcon
         href={ROUTER_PATH.MY_PAGE}
         icon={<UserIcon />}
         label={PATH_NAME.MY_PAGE}
@@ -24,4 +24,4 @@ const NavIcon = () => {
   );
 };
 
-export default NavIcon;
+export default HeaderNavigation;
